requirements.txt
selenium
cryptography
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.4
django-cors-headers==3.5.0
idna==2.10
pytz==2020.4
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.2
beautifulsoup4
exchangelib
django_apscheduler
djangorestframework==3.12.4